+++ examples/data-sources/cpanel_apache_handler/data-source.tf
@@ -1,0 +1,3 @@
+data "cpanel_apache_handler" "example" {
+  extension = ".example-handler"
+}