+++ examples/resources/cpanel_apache_handler/resource.tf
@@ -1,0 +1,4 @@
+resource "cpanel_apache_handler" "example" {
+  extension = ".example-handler"
+  handler   = "example-handler"
+}