+++ internal/provider/provider.go
@@ -11,6 +11,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"os"
 	"terraform-provider-cpanel/internal/cpanel"
+	"terraform-provider-cpanel/internal/cpanel/apachehandler"
 	"terraform-provider-cpanel/internal/cpanel/apitoken"
 	"terraform-provider-cpanel/internal/cpanel/cron"
 	"terraform-provider-cpanel/internal/cpanel/ddns"
@@ -56,8 +57,8 @@ func (p *cpanelProvider) Metadata(_ context.Context, _ provider.MetadataRequest,
 // Schema defines the provider-level schema for configuration data.
 func (p *cpanelProvider) Schema(_ context.Context, _ provider.SchemaRequest, resp *provider.SchemaResponse) {
 	resp.Schema = schema.Schema{
-		Description:         "Manage cPanel account API tokens, cron jobs, DNS records, Dynamic DNS domains, web domains, HTTP redirects, custom MIME types, email accounts, forwarders, autoresponders, FTP accounts, website IP blocks, and MySQL, MariaDB, and PostgreSQL users and databases.",
-		MarkdownDescription: "Manage cPanel account API tokens, cron jobs, DNS records, Dynamic DNS domains, web domains, HTTP redirects, custom MIME types, email accounts, forwarders, autoresponders, FTP accounts, website IP blocks, and MySQL, MariaDB, and PostgreSQL users and databases.",
+		Description:         "Manage cPanel account API tokens, cron jobs, DNS records, Dynamic DNS domains, web domains, HTTP redirects, custom MIME types and Apache handlers, email accounts, forwarders, autoresponders, FTP accounts, website IP blocks, and MySQL, MariaDB, and PostgreSQL users and databases.",
+		MarkdownDescription: "Manage cPanel account API tokens, cron jobs, DNS records, Dynamic DNS domains, web domains, HTTP redirects, custom MIME types and Apache handlers, email accounts, forwarders, autoresponders, FTP accounts, website IP blocks, and MySQL, MariaDB, and PostgreSQL users and databases.",
 		Attributes: map[string]schema.Attribute{
 			"username": schema.StringAttribute{
 				Optional:            true,
@@ -195,6 +196,7 @@ func (p *cpanelProvider) Configure(ctx context.Context, req provider.ConfigureRe
 	}
 
 	// Initialize module clients
+	apacheHandlerClient := apachehandler.NewClient(client)
 	apiTokenClient := apitoken.NewClient(client)
 	cronClient := cron.NewClient(client)
 	dnsClient := cpaneldns.NewClient(client)
@@ -211,32 +213,34 @@ func (p *cpanelProvider) Configure(ctx context.Context, req provider.ConfigureRe
 	// Make the module clients available during DataSource and Resource
 	// type Configure methods.
 	resp.DataSourceData = map[string]interface{}{
-		"apitoken":   apiTokenClient,
-		"cron":       cronClient,
-		"dns":        dnsClient,
-		"domain":     domainClient,
-		"ddns":       dynamicDNSClient,
-		"email":      emailClient,
-		"ftp":        ftpClient,
-		"ipblock":    ipBlockClient,
-		"mimetype":   mimeTypeClient,
-		"mysql":      mySQLClient,
-		"postgresql": postgreSQLClient,
-		"redirect":   redirectClient,
+		"apachehandler": apacheHandlerClient,
+		"apitoken":      apiTokenClient,
+		"cron":          cronClient,
+		"dns":           dnsClient,
+		"domain":        domainClient,
+		"ddns":          dynamicDNSClient,
+		"email":         emailClient,
+		"ftp":           ftpClient,
+		"ipblock":       ipBlockClient,
+		"mimetype":      mimeTypeClient,
+		"mysql":         mySQLClient,
+		"postgresql":    postgreSQLClient,
+		"redirect":      redirectClient,
 	}
 	resp.ResourceData = map[string]interface{}{
-		"apitoken":   apiTokenClient,
-		"cron":       cronClient,
-		"dns":        dnsClient,
-		"domain":     domainClient,
-		"ddns":       dynamicDNSClient,
-		"email":      emailClient,
-		"ftp":        ftpClient,
-		"ipblock":    ipBlockClient,
-		"mimetype":   mimeTypeClient,
-		"mysql":      mySQLClient,
-		"postgresql": postgreSQLClient,
-		"redirect":   redirectClient,
+		"apachehandler": apacheHandlerClient,
+		"apitoken":      apiTokenClient,
+		"cron":          cronClient,
+		"dns":           dnsClient,
+		"domain":        domainClient,
+		"ddns":          dynamicDNSClient,
+		"email":         emailClient,
+		"ftp":           ftpClient,
+		"ipblock":       ipBlockClient,
+		"mimetype":      mimeTypeClient,
+		"mysql":         mySQLClient,
+		"postgresql":    postgreSQLClient,
+		"redirect":      redirectClient,
 	}
 
 	tflog.Info(ctx, "Configured module clients", map[string]any{"success": true})
@@ -246,6 +250,7 @@ func (p *cpanelProvider) Configure(ctx context.Context, req provider.ConfigureRe
 func (p *cpanelProvider) DataSources(_ context.Context) []func() datasource.DataSource {
 	return []func() datasource.DataSource{
 		NewAddonDomainDataSource,
+		NewApacheHandlerDataSource,
 		NewAPITokenDataSource,
 		NewCronJobDataSource,
 		NewDNSRecordDataSource,
@@ -271,6 +276,7 @@ func (p *cpanelProvider) DataSources(_ context.Context) []func() datasource.Data
 func (p *cpanelProvider) Resources(_ context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		NewAddonDomainResource,
+		NewApacheHandlerResource,
 		NewAPITokenResource,
 		NewCronJobResource,
 		NewDNSRecordResource,